* Move serial handlers into better namespace.
