+++ src/panoptes/utils/serial/handlers/protocol_no_op.py
